Moving segrep specific tests to SegmentReplicationIndexShardTests.
