+++ server/src/test/java/org/opensearch/index/shard/IndexShardTests.java
@@ -206,10 +206,7 @@ import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
 import static org.hamcrest.Matchers.oneOf;
 import static org.hamcrest.Matchers.sameInstance;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 import static org.opensearch.cluster.routing.TestShardRouting.newShardRouting;
 import static org.opensearch.common.lucene.Lucene.cleanLuceneIndex;
 import static org.opensearch.common.xcontent.ToXContent.EMPTY_PARAMS;
@@ -3521,42 +3518,6 @@ public class IndexShardTests extends IndexShardTestCase {
         IndexShard shard = newStartedShard(p -> newShard(null), true);
         List<ReferenceManager.RefreshListener> refreshListeners = shard.getEngine().config().getInternalRefreshListener();
         assertFalse(refreshListeners.stream().anyMatch(e -> e instanceof CheckpointRefreshListener));
-        closeShards(shard);
-    }
-
-    /**
-     * here we are starting a new primary shard in PrimaryMode and testing if the shard publishes checkpoint after refresh.
-     */
-    public void testPublishCheckpointOnPrimaryMode() throws IOException {
-        final SegmentReplicationCheckpointPublisher mock = mock(SegmentReplicationCheckpointPublisher.class);
-        IndexShard shard = newStartedShard(p -> newShard(mock), true);
-        CheckpointRefreshListener refreshListener = new CheckpointRefreshListener(shard, mock);
-        refreshListener.afterRefresh(true);
-
-        // verify checkpoint is published
-        verify(mock, times(1)).publish(any());
-        closeShards(shard);
-    }
-
-    /**
-     * here we are starting a new primary shard in PrimaryMode intially and starting relocation handoff. Later we complete relocation handoff then shard is no longer
-     * in PrimaryMode, and we test if the shard does not publish checkpoint after refresh.
-     */
-    public void testPublishCheckpointAfterRelocationHandOff() throws IOException {
-        final SegmentReplicationCheckpointPublisher mock = mock(SegmentReplicationCheckpointPublisher.class);
-        IndexShard shard = newStartedShard(p -> newShard(mock), true);
-        CheckpointRefreshListener refreshListener = new CheckpointRefreshListener(shard, mock);
-        String id = shard.routingEntry().allocationId().getId();
-
-        // Starting relocation handoff
-        shard.getReplicationTracker().startRelocationHandoff(id);
-
-        // Completing relocation handoff
-        shard.getReplicationTracker().completeRelocationHandoff();
-        refreshListener.afterRefresh(true);
-
-        // verify checkpoint is not published
-        verify(mock, times(0)).publish(any());
         closeShards(shard);
     }
